Add _id, _block_range columns, exclude constraint, smart tags, hooks

+++ packages/node/src/indexer/store.service.ts
@@ -6,11 +6,19 @@ import { Injectable } from '@nestjs/common';
 import { hexToU8a, u8aToBuffer } from '@polkadot/util';
 import { blake2AsHex } from '@polkadot/util-crypto';
 import { Entity, Store } from '@subql/types';
-import { GraphQLModelsRelationsEnums } from '@subql/utils';
+import {
+  GraphQLModelsRelationsEnums,
+  GraphQLRelationsType,
+} from '@subql/utils';
 import { camelCase, flatten, isEqual, upperFirst } from 'lodash';
 import {
   CreationAttributes,
+  DataTypes,
   Model,
+  ModelAttributeColumnOptions,
+  ModelAttributes,
+  ModelStatic,
+  Op,
   QueryTypes,
   Sequelize,
   Transaction,
@@ -22,6 +30,7 @@ import { modelsTypeToModelAttributes } from '../utils/graphql';
 import { getLogger } from '../utils/logger';
 import { camelCaseObjectKey } from '../utils/object';
 import {
+  commentTableQuery,
   commentConstraintQuery,
   createNotifyTrigger,
   createSendNotificationTriggerFunction,
@@ -29,7 +38,11 @@ import {
   dropNotifyTrigger,
   getFkConstraint,
   getNotifyTriggers,
+  SmartTags,
   smartTags,
+  getVirtualFkTag,
+  addTagsToForeignKeyMap,
+  createExcludeConstraintQuery,
 } from '../utils/sync-helper';
 import { getYargsOption } from '../yargs';
 import {
@@ -67,6 +80,8 @@ export class StoreService {
   private poiRepo: PoiRepo;
   private metaDataRepo: MetadataRepo;
   private operationStack: StoreOperations;
+  private blockHeight: number;
+  private historical: boolean;
 
   constructor(
     private sequelize: Sequelize,
@@ -80,6 +95,7 @@ export class StoreService {
   ): Promise<void> {
     this.schema = schema;
     this.modelsRelations = modelsRelations;
+    this.historical = argv['experimental-historical'];
     try {
       await this.syncSchema(this.schema);
     } catch (e) {
@@ -161,6 +177,9 @@ export class StoreService {
       if (indexes.length > this.config.indexCountLimit) {
         throw new Error(`too many indexes on entity ${model.name}`);
       }
+      if (this.historical) {
+        this.addIdAndBlockRangeAttributes(attributes);
+      }
       const sequelizeModel = this.sequelize.define(model.name, attributes, {
         underscored: true,
         comment: model.description,
@@ -170,6 +189,12 @@ export class StoreService {
         schema,
         indexes,
       });
+      if (this.historical) {
+        this.addScopeAndBlockHeightHooks(sequelizeModel);
+        extraQueries.push(
+          ...createExcludeConstraintQuery(schema, sequelizeModel.tableName),
+        );
+      }
       if (argv.subscription) {
         const triggerName = `${schema}_${sequelizeModel.tableName}_notify_trigger`;
         const triggers = await this.sequelize.query(getNotifyTriggers(), {
@@ -191,9 +216,14 @@ export class StoreService {
         extraQueries.push(dropNotifyTrigger(schema, sequelizeModel.tableName));
       }
     }
+    const foreignKeyMap = new Map<string, Map<string, SmartTags>>();
     for (const relation of this.modelsRelations.relations) {
       const model = this.sequelize.model(relation.from);
       const relatedModel = this.sequelize.model(relation.to);
+      if (this.historical) {
+        this.addRelationToMap(relation, foreignKeyMap, model, relatedModel);
+        continue;
+      }
       switch (relation.type) {
         case 'belongsTo': {
           model.belongsTo(relatedModel, { foreignKey: relation.foreignKey });
@@ -249,6 +279,13 @@ export class StoreService {
           throw new Error('Relation type is not supported');
       }
     }
+    foreignKeyMap.forEach((keys, tableName) => {
+      const comment = Array.from(keys.values())
+        .map((tags) => smartTags(tags, '|'))
+        .join('\n');
+      const query = commentTableQuery(`"${schema}"."${tableName}"`, comment);
+      extraQueries.push(query);
+    });
     if (this.config.proofOfIndex) {
       this.poiRepo = PoiFactory(this.sequelize, schema);
     }
@@ -258,6 +295,91 @@ export class StoreService {
     for (const query of extraQueries) {
       await this.sequelize.query(query);
     }
+  }
+
+  private addRelationToMap(
+    relation: GraphQLRelationsType,
+    foreignKeys: Map<string, Map<string, SmartTags>>,
+    model: ModelStatic<any>,
+    relatedModel: ModelStatic<any>,
+  ) {
+    switch (relation.type) {
+      case 'belongsTo': {
+        addTagsToForeignKeyMap(
+          foreignKeys,
+          model.tableName,
+          relation.foreignKey,
+          {
+            foreignKey: getVirtualFkTag(
+              relation.foreignKey,
+              relatedModel.tableName,
+            ),
+          },
+        );
+        break;
+      }
+      case 'hasOne': {
+        addTagsToForeignKeyMap(
+          foreignKeys,
+          relatedModel.tableName,
+          relation.foreignKey,
+          {
+            singleForeignFieldName: relation.fieldName,
+          },
+        );
+        break;
+      }
+      case 'hasMany': {
+        addTagsToForeignKeyMap(
+          foreignKeys,
+          relatedModel.tableName,
+          relation.foreignKey,
+          {
+            foreignFieldName: relation.fieldName,
+          },
+        );
+        break;
+      }
+      default:
+        throw new Error('Relation type is not supported');
+    }
+  }
+
+  private addIdAndBlockRangeAttributes(
+    attributes: ModelAttributes<Model<any, any>, any>,
+  ): void {
+    (attributes.id as ModelAttributeColumnOptions).primaryKey = false;
+    attributes.__id = {
+      type: DataTypes.UUID,
+      defaultValue: DataTypes.UUIDV4,
+      allowNull: false,
+      primaryKey: true,
+    } as ModelAttributeColumnOptions;
+    attributes.__block_range = {
+      type: DataTypes.RANGE(DataTypes.BIGINT),
+      allowNull: false,
+    } as ModelAttributeColumnOptions;
+  }
+
+  private addScopeAndBlockHeightHooks(sequelizeModel: ModelStatic<any>): void {
+    sequelizeModel.addScope('defaultScope', {
+      attributes: {
+        exclude: ['__id', '__block_range'],
+      },
+    });
+    sequelizeModel.addHook('beforeFind', (options) => {
+      options.where.__block_range = {
+        [Op.contains]: this.blockHeight as any,
+      };
+    });
+    sequelizeModel.addHook('beforeValidate', (attributes, options) => {
+      attributes.__block_range = [this.blockHeight, null];
+    });
+    sequelizeModel.addHook('beforeBulkCreate', (instances, options) => {
+      instances.forEach((item) => {
+        item.__block_range = [this.blockHeight, null];
+      });
+    });
   }
 
   validateNotifyTriggers(
@@ -288,6 +410,10 @@ export class StoreService {
     if (this.config.proofOfIndex) {
       this.operationStack = new StoreOperations(this.modelsRelations.models);
     }
+  }
+
+  setBlockHeight(blockHeight: number): void {
+    this.blockHeight = blockHeight;
   }
 
   async setMetadataBatch(
@@ -379,6 +505,28 @@ group by
     return rows.map((result) => camelCaseObjectKey(result)) as IndexField[];
   }
 
+  private async markAsDeleted(model: ModelStatic<any>, id: string) {
+    return model.update(
+      {
+        __block_range: this.sequelize.fn(
+          'int8range',
+          this.sequelize.fn('lower', this.sequelize.col('_block_range')),
+          this.blockHeight,
+        ),
+      },
+      {
+        hooks: false,
+        transaction: this.tx,
+        where: {
+          id: id,
+          __block_range: {
+            [Op.contains]: this.blockHeight as any,
+          },
+        },
+      },
+    );
+  }
+
   getStore(): Store {
     return {
       get: async (entity: string, id: string): Promise<Entity | undefined> => {
@@ -441,9 +589,16 @@ group by
       set: async (entity: string, _id: string, data: Entity): Promise<void> => {
         const model = this.sequelize.model(entity);
         assert(model, `model ${entity} not exists`);
-        await model.upsert(data as unknown as CreationAttributes<Model>, {
-          transaction: this.tx,
-        });
+        if (this.historical) {
+          await this.markAsDeleted(model, data.id);
+          await model.create(data as unknown as CreationAttributes<Model>, {
+            transaction: this.tx,
+          });
+        } else {
+          await model.upsert(data as unknown as CreationAttributes<Model>, {
+            transaction: this.tx,
+          });
+        }
         if (this.config.proofOfIndex) {
           this.operationStack.put(OperationType.Set, entity, data);
         }
@@ -463,7 +618,11 @@ group by
       remove: async (entity: string, id: string): Promise<void> => {
         const model = this.sequelize.model(entity);
         assert(model, `model ${entity} not exists`);
-        await model.destroy({ where: { id }, transaction: this.tx });
+        if (this.historical) {
+          await this.markAsDeleted(model, id);
+        } else {
+          await model.destroy({ where: { id }, transaction: this.tx });
+        }
         if (this.config.proofOfIndex) {
           this.operationStack.put(OperationType.Remove, entity, id);
         }
